refactor(cli): Rename "container unpack" command to "package unpack"
More sensible naming.
The old command is kept for now , returning an error message that it was
renamed.
